+++ sentry-ruby/lib/sentry/span.rb
@@ -68,13 +68,14 @@ module Sentry
       op: nil,
       status: nil,
       trace_id: nil,
+      span_id: nil,
       parent_span_id: nil,
       sampled: nil,
       start_timestamp: nil,
       timestamp: nil
     )
       @trace_id = trace_id || SecureRandom.uuid.delete("-")
-      @span_id = SecureRandom.hex(8)
+      @span_id = span_id || SecureRandom.hex(8)
       @parent_span_id = parent_span_id
       @sampled = sampled
       @start_timestamp = start_timestamp || Sentry.utc_now.to_f