+++ sentry-opentelemetry/lib/sentry-opentelemetry.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "sentry-ruby"
+require "opentelemetry-sdk"
+
+require "sentry/opentelemetry/version"
+require "sentry/opentelemetry/span_processor"
+require "sentry/opentelemetry/propagator"