+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Features
+
+- Add OpenTelemetry support with new `sentry-opentelemetry` gem
+  - Add `config.instrumenter` to switch between sentry and otel instrumentation [#1944](https://github.com/getsentry/sentry-ruby/pull/1944)
+
 ## 5.6.0
 
 ### Features